remove the need for ECReferenceTypesCache in link table relationships, fix a weird type constraint error
