+++ core/backend/src/core-backend.ts
@@ -12,8 +12,6 @@ export * from "./CodeSpecs";
 export * from "./CloudSqlite";
 export * from "./ConcreteEntityId";
 export * from "./EntityReference";
-export * from "./ECReferenceTypesCache";
-export * from "./EntityUnifier";
 export * from "./DisplayStyle";
 export * from "./ECDb";
 export * from "./ECSchemaXmlContext";
@@ -49,6 +47,7 @@ export * from "./domains/FunctionalElements";
 export * from "./domains/GenericSchema";
 export * from "./domains/GenericElements";
 export { IModelJsNative, NativeCloudSqlite, NativeLoggerCategory } from "@bentley/imodeljs-native";
+export * from "./IModelElementCloneContext";
 export * from "./IModelCloneContext";
 export * from "./IModelHost";
 export * from "./IpcHost";
