+++ core/transformer/src/IModelTransformer.ts
@@ -15,7 +15,7 @@ import * as ECSchemaMetaData from "@itwin/ecschema-metadata";
 import { Point3d, Transform } from "@itwin/core-geometry";
 import {
   ChannelRootAspect, ChangeSummaryManager, ConcreteEntity
-  DefinitionElement, DefinitionModel, DefinitionPartition, ECReferenceTypesCache, ECSqlStatement, Element, ElementAspect, ElementMultiAspect,
+  DefinitionElement, DefinitionModel, DefinitionPartition, ECSqlStatement, Element, ElementAspect, ElementMultiAspect,
   ElementOwnsExternalSourceAspects, ElementRefersToElements, ElementUniqueAspect, Entity, EntityReferences, EntityUnifier, ExternalSource, ExternalSourceAspect, ExternalSourceAttachment,
   FolderLink, GeometricElement2d, GeometricElement3d, IModelCloneContext, IModelDb, IModelHost, IModelJsFs, InformationPartitionElement, KnownLocations, Model,
   RecipeDefinitionElement, Relationship, RelationshipProps, Schema, SQLiteDb, Subject, SynchronizationConfigLink,
@@ -30,6 +30,8 @@ import { IModelImporter, IModelImporterState, OptimizeGeometryOptions } from "./
 import { TransformerLoggerCategory } from "./TransformerLoggerCategory";
 import { PendingReference, PendingReferenceMap } from "./PendingReferenceMap";
 import { EntityMap } from "./EntityMap";
+import { IModelCloneContext } from "./IModelCloneContext";
+import { EntityUnifier } from "./EntityUnifier";
 
 const loggerCategory: string = TransformerLoggerCategory.IModelTransformer;
 
@@ -630,7 +632,7 @@ export class IModelTransformer extends IModelExportHandler {
     const missingReferences = new EntityReferenceSet();
     let thisPartialElem: PartiallyCommittedEntity | undefined;
 
-    for (const referenceId of entity.getReferenceConcreteIds(ECReferenceTypesCache.globalCache)) {
+    for (const referenceId of entity.getReferenceConcreteIds()) {
       // TODO: probably need to rename from 'id' to 'ref' so these names aren't so ambiguous
       const referenceIdInTarget = this.context.findTargetEntityId(referenceId);
       const alreadyImported = EntityReferences.isValid(referenceIdInTarget);
